add startViewLoadSpan as a platform extenison
